Corrección tipo ID pokemon

+++ backend/pokemonbackend/src/main/java/com/pokemon/pokemonbackend/dto/PokemonResponseDTO.java
@@ -1,14 +1,14 @@
 package com.pokemon.pokemonbackend.dto;
 
 public class PokemonResponseDTO {
-    private Long id;
+    private int id;
     private int pokedexNumber;
     private String name;
     private String primaryType;
     private String secondaryType;
 
     public PokemonResponseDTO(
-            Long id,
+            int id,
             int pokedexNumber,
             String name,
             String primaryType,
@@ -21,7 +21,7 @@ public class PokemonResponseDTO {
         this.secondaryType = secondaryType;
     }
 
-    public Long getId() { return id; }
+    public int getId() { return id; }
     public int getPokedexNumber() { return pokedexNumber; }
     public String getName() { return name; }
     public String getPrimaryType() { return primaryType; }
